Changed problem Pound To Dollars

+++ Prgramming Fundamentals With Java/L02DataTypesandVariablesLab/P02PoundsToDollars.java
@@ -6,7 +6,7 @@ public class P02PoundsToDollars {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int pounds = Integer.parseInt(scanner.nextLine());
+        double pounds = Double.parseDouble(scanner.nextLine());
 
         double poundsToDollars = pounds * 1.36;
         System.out.printf("%.3f%n", poundsToDollars);
